Fix bunyan logger
One of the first problems with the library was the bad implementation of this logger library. The problem was the first approach but now it's correct.The name of the log files change if it changes on config file. So now is configurable. The path of the log files now is not posible to change on config files.

[00000]