Add verification JWT to config
Add new value to config.json to add doble check verification o Redis and MongoDB if this value is set to true, if not, will be check only on redis. The JWT will bet set when user signin on the system. If you check to true will be more slowly but more secure. This part is not finish yet, will be soon.