Add token checker functionality
Add the functionality and their test, this middleware check if the token exist on redis or not, if not will thrown an Error on response. A future will be to check the integrity of the token, but probably is better to do it on other middleware and only optional by config.js(Just for choose if the user want a rapid check or a secure check).