Add new exceptions to JWT module
New exceptions SIGNATURE_VERIFICATION, EXPIRED_TOKEN, TOKEN_NOT_ACTIVE, ALGORITHM_NOT_SUPPORTED and TOKEN_NOT_SUPPLIED on JWT module when jet simple module encode or decode the token. It is necessary for the middleware token checker to control all posible exceptions when the user send it. The test files should be finish in the next commit.