Improve JWT test
Add new errors to JWT module, these errors comes from jwt-simple module that throw some exceptions but the app need to add them to the error manager to have a better management of the errors. Also move fake.json of the errors to mocks folder, before that was on config folder.